requirements.txt
discord.py==1.7.3
Pillow==8.4.0